clear task-manager alarms
desc:clear task-manager alarms

issue:https://gitee.com/openharmony/commonlibrary_ets_utils/issues/I8Y5YN

+++ js_concurrent_module/taskpool/task_manager.cpp
@@ -91,9 +91,9 @@ TaskManager::~TaskManager()
 void TaskManager::CountTraceForWorker()
 {
     std::lock_guard<std::recursive_mutex> lock(workersMutex_);
-    int64_t threadNum = workers_.size();
-    int64_t idleWorkers = idleWorkers_.size();
-    int64_t timeoutWorkers = timeoutWorkers_.size();
+    int64_t threadNum = static_cast<int64_t>(workers_.size());
+    int64_t idleWorkers = static_cast<int64_t>(idleWorkers_.size());
+    int64_t timeoutWorkers = static_cast<int64_t>(timeoutWorkers_.size());
     HITRACE_HELPER_COUNT_TRACE("timeoutThreadNum", timeoutWorkers);
     HITRACE_HELPER_COUNT_TRACE("threadNum", threadNum);
     HITRACE_HELPER_COUNT_TRACE("runningThreadNum", threadNum - idleWorkers);
